main.py
import argparse
import cv2
import os
from PIL import Image
import numpy as np
import tensorflow as tf
from utils.image_utils import annotate_weeds, annotate_diseases
from utils.moisture_utils import estimate_moisture
from utils.report_utils import generate_report

def load_image(image_path):
    """Load image from path."""
    img = Image.open(image_path)
    return np.array(img)

def detect_weeds(image, model_path):
    """Detect weeds using a pre-trained model."""
    # Placeholder: Assume model is YOLO-like, but for demo, return dummy boxes
    # In real impl, load model and predict
    # model = tf.keras.models.load_model(model_path)
    # boxes = model.predict(image)
    boxes = [[100, 100, 200, 200], [300, 300, 400, 400]]  # Dummy
    return {'count': len(boxes), 'boxes': boxes}

def detect_diseases(image, model_path):
    """Detect diseases using a segmentation model."""
    # Placeholder: Assume segmentation model
    # model = tf.keras.models.load_model(model_path)
    # mask = model.predict(image)
    mask = np.zeros((image.shape[0], image.shape[1]), dtype=np.uint8)  # Dummy mask
    mask[50:150, 50:150] = 255  # Dummy diseased area
    return {'mask': mask, 'type': 'Powdery Mildew'}  # Dummy type

def main():
    parser = argparse.ArgumentParser(description='Green Roof Monitoring System')
    parser.add_argument('--drone_image', type=str, help='Path to drone RGB image for weed detection')
    parser.add_argument('--leaf_image', type=str, help='Path to leaf image for disease detection')
    parser.add_argument('--soil_image', type=str, help='Path to soil image for moisture estimation')
    parser.add_argument('--moisture_sensor', type=float, help='Soil moisture sensor percentage (0-100)')
    parser.add_argument('--temperature', type=float, required=True, help='Temperature in Celsius')
    parser.add_argument('--humidity', type=float, required=True, help='Humidity percentage')
    parser.add_argument('--output_dir', type=str, default='output', help='Output directory for annotated images and report')
    
    args = parser.parse_args()
    
    # Load images
    drone_img = load_image(args.drone_image) if args.drone_image else None
    leaf_img = load_image(args.leaf_image) if args.leaf_image else None
    soil_img = load_image(args.soil_image) if args.soil_image else None
    
    # Weed detection
    weed_data = {}
    if drone_img is not None:
        weed_data = detect_weeds(drone_img, 'models/weed_detector.h5')
        annotated_drone = annotate_weeds(drone_img, weed_data['boxes'])
        cv2.imwrite(os.path.join(args.output_dir, 'annotated_drone.jpg'), cv2.cvtColor(annotated_drone, cv2.COLOR_RGB2BGR))
    
    # Disease detection
    disease_data = {}
    if leaf_img is not None:
        disease_result = detect_diseases(leaf_img, 'models/disease_model.h5')
        annotated_leaf, percentage = annotate_diseases(leaf_img, disease_result['mask'])
        disease_data = {'percentage': percentage, 'type': disease_result['type']}
        cv2.imwrite(os.path.join(args.output_dir, 'annotated_leaf.jpg'), cv2.cvtColor(annotated_leaf, cv2.COLOR_RGB2BGR))
    
    # Moisture estimation
    moisture_level = 'Unknown'
    if soil_img is not None:
        moisture_level = estimate_moisture(soil_img, args.moisture_sensor)
    
    # Generate report
    report_path = generate_report(weed_data, disease_data, moisture_level, args.temperature, args.humidity, args.output_dir)
    
    print(f'Report generated at: {report_path}')

if __name__ == '__main__':
    main()
