requirements.txt
opencv-python==4.8.1.78
tensorflow==2.13.0
pillow==10.0.1
numpy==1.24.3
flask==2.3.3
werkzeug==2.3.7
